workflows/react_distributed/react_orchestrator/requirements.txt
bedrock-agentcore
boto3>=1.40.0
langgraph
python-dotenv
